fix crash screen

+++ include/n64/PR/ultratypes.h
@@ -41,7 +41,7 @@ typedef signed int             s32;
 typedef unsigned int           u32;
 
 typedef float  f32;
-typedef float f64;
+typedef double f64;
 
 #ifdef TARGET_N64
 typedef signed long long int   s64;
